refactor(strategies): split monolithic file into focused modules
Split requirement_strategies.py (574 lines) into 6 focused modules following
single responsibility principle and eliminating circular dependencies.

New architecture:
- strategy_utils.py: Shared logging/response utilities
- base_strategy.py: Abstract RequirementStrategy base class
- blocking_strategy.py: Manual satisfaction strategy
- dynamic_strategy.py: Calculator-based threshold strategy
- guard_strategy.py: Boolean condition guards
- strategy_registry.py: Central STRATEGIES registry

Benefits:
- Clean acyclic dependency graph (utils → base → strategies → registry)
- Each module has single clear responsibility
- Easier to test, maintain, and extend
- Better aligns with Strategy pattern (ADR-004)

Testing: All 539 tests pass with zero regressions
